# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/admin/users:
    get:
      operationId: UsersAdminController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 20
            type: number
        - name: role
          required: false
          in: query
          schema:
            type: string
            enum:
              - ADMIN
              - STAFF
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersAdminResponseDto'
      security: &ref_17
        - bearer: []
      summary: List admin panel users (ADMIN/STAFF)
      tags: &ref_18
        - admin-users
    post:
      operationId: UsersAdminController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserAdminDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAdminResponseDto'
      security: *ref_17
      summary: Create admin panel user
      tags: *ref_18
  /api/v1/admin/users/{id}:
    patch:
      operationId: UsersAdminController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserAdminDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAdminResponseDto'
      security: *ref_17
      summary: Update role and/or password
      tags: *ref_18
...
info:
  title: Dog Rescue API
  description: REST API for Dog Rescue (MVP scaffold)
  version: '1.0'
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UserAdminResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: staff@dogerescue.org
        role:
          type: string
          enum:
            - ADMIN
            - STAFF
            - USER
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - email
        - role
        - createdAt
        - updatedAt
    PaginatedUsersAdminResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserAdminResponseDto'
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - items
        - total
        - page
        - limit
    CreateUserAdminDto:
      type: object
      properties:
        email:
          type: string
          example: staff@dogerescue.org
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum:
            - ADMIN
            - STAFF
      required:
        - email
        - password
        - role
    UpdateUserAdminDto:
      type: object
      properties:
        role:
          type: string
          enum:
            - ADMIN
            - STAFF
        password:
          type: string
          minLength: 8
...